% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_radial.R
\docType{data}
\name{data_radial}
\alias{data_radial}
\title{Two Sample Summary GWAS data published in Do et al (2014).}
\format{
A data frame with 185 rows and 21 variables. Specifically this includes the following information:
\describe{
\item{\code{rsid}} {The identification number for each variant}
\item{\code{a1}} {The reference allele for each variant}
\item{\code{a2}} {The other allele for each variant}
\item{\code{chr}} {The chromosome number in which each variant is located}
\item{\code{pos}} {The genomic position for the genetic variant relative to chromosome number}
\item{\code{ldlcbeta}} {The association estimate for the genetic variant obtained by regressing LDL-C upon the genetic variant}
\item{\code{hdlcbeta}} {The association estimate obtained by regressing HDL-C upon the genetic variant}
\item{\code{tgbeta}} {The association estimate obtained by regressing triglycerides upon the genetic variant}
\item{\code{chdbeta}} {The association estimate for CHD obtained by regressing CHD upon the genetic variant}
\item{\code{ldlcp2}} {The p-value corresponding to association estimate \code{ldlcbeta}}
\item{\code{hdlcp2}} {The p-value corresponding to association estimate \code{hdlcbeta}}
\item{\code{tgp2}} {The p-value corresponding to association estimate \code{tgbeta}}
\item{\code{chdp2}} {The p-value corresponding to association estimate \code{chdbeta}}
\item{\code{ldlcz2}} {The z-score corresponding to association estimate \code{ldlcbeta}}
\item{\code{ldlcse}} {The standard error corresponding to association estimate \code{ldlcbeta}}
\item{\code{hdlcz2}} {The z-score corresponding to association estimate \code{hdlcbeta}}
\item{\code{hdlcse}} {The standard error corresponding to association estimate \code{hdlcbeta}}
\item{\code{tgz}} {The z-score corresponding to association estimate \code{tgbeta}}
\item{\code{tgse}} {The standard error corresponding to association estimate \code{tgbeta}}
\item{\code{chdz}} {The z-score corresponding to association estimate \code{chdbeta}}
\item{\code{chdse}} {The standard error corresponding to association estimate \code{chdbeta}}

}
}
\source{
\itemize{
\item \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3904346/}
\item \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3838666/}
\item \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3119261/}
}
}
\usage{
data_radial
}
\description{
A dataset containing summary data GWAS data for 185 independent SNPs with respect to lipid fractions and coronary heart disease (CHD), previously published in Do et al (2013).
Lipid fractions include low-density lipoprotein (LDL-C),high-density lipoprotein (HDL-C), and triglycerides, with association estimates
obtained from the Global Lipids Genetics Consortium (GLGC) (Willer et al, 2013). Summary data for CHD was obtained from the CARDIoGRAM study (Schunkert et al, 2011).
Association estimates were obtained by regressing each phenotype upon the genetic variant and additional adjusted covariates, with links to further information
related to each study presented below.
}
\details{
data_radial
}
\examples{

head(data_radial)
}
\author{
Wes Spiller; Jack Bowden; Tom Palmer.
}
\keyword{datasets}
