% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funnel_radial.R
\name{funnel_radial}
\alias{funnel_radial}
\title{funnel_radial}
\usage{
funnel_radial(r_object, show_transform)
}
\arguments{
\item{r_object}{An object of class \code{"IVW"} or \code{"egger"}. For visualising both estimates simultaneously, both objects should be included as a vector c(\code{A},\code{B}), where \code{A} and \code{B} denote the \code{"IVW"} and \code{"egger"} objects respectively.}

\item{show_transform}{Indicates whether to produce a plot including lines showing the magnitude of the MR Egger transformation for each variant (\code{TRUE}), or a scatterplot showing only the variants and corresponding estimates (\code{FALSE}).}
}
\value{
A ggplot object containing a radial funnel plot of either the IVW, MR-Egger, or both estimates simultaneously.
}
\description{
A function for producing generalized radial IVW and MR-Egger funnel plots either individually or simultaneously. The function also allows for lines indicating the magnitude for the MR Egger transformation for each variant, though it should be noted that this distance is a function of the weight attributed to the variant, and is therefore not indicative of outliers.
}
\examples{

funnel_radial(r_object,TRUE) 
}
\references{
Bowden, J. et al., 2017. "Improving the visualisation, interpretation and analysis of two-sample summary data Mendelian randomization via the radial plot and radial regression." BioRxiv. doi: https://doi.org/10.1101/200378. Available at: \url{https://www.biorxiv.org/content/early/2017/10/11/200378}
}
\author{
Wes Spiller; Jack Bowden.
}
