% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrinput_to_rmr_format.R
\name{mrinput_to_rmr_format}
\alias{mrinput_to_rmr_format}
\title{Convert an object of class MRInput from the MendelianRandomization package to the RadialMR rmr_format class}
\usage{
mrinput_to_rmr_format(dat)
}
\arguments{
\item{dat}{Object from \code{\link[MendelianRandomization:mr_input]{MendelianRandomization::mr_input}}.}
}
\value{
Object of class \code{rmr_format}, the RadialMR format
}
\description{
Creates a data.frame with class \code{rmr_format} from an object of class \code{MRInput} generated by \code{\link[MendelianRandomization:mr_input]{MendelianRandomization::mr_input}}.
}
\examples{
if (require("MendelianRandomization", quietly = TRUE)) {
dat <- data_radial[data_radial[,10] < 5e-8,]
dat <- MendelianRandomization::mr_input(bx = dat$ldlcbeta,
                                        bxse = dat$ldlcse,
                                        by = dat$chdbeta,
                                        byse = dat$chdse,
                                        snps = dat$rsid)
dat <- mrinput_to_rmr_format(dat)
head(dat)
class(dat)
}
}
