% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_radial.R
\name{plotly_radial}
\alias{plotly_radial}
\title{plotly_radial}
\usage{
plotly_radial(r_object, TEST)
}
\arguments{
\item{r_object}{An object of class \code{"IVW"} or \code{"egger"}.}

\item{TEST}{Logical; indicating whether testing the function}
}
\value{
A plotly object containing a radial plot of either the IVW or MR-Egger estimates. Hovering the mouse over individual datapoints will highlight the corresponding SNP identification number for that observation.
}
\description{
A function for producing interactive radial IVW and MR-Egger plots individually.The function utilises the output from the \code{IVW_radial} and \code{egger_radial} functions.
}
\examples{
ldl.dat <- data_radial[data_radial[,10]<5*10^-8,]
ldl.fdat <- format_radial(ldl.dat[,6], ldl.dat[,9],
                          ldl.dat[,15], ldl.dat[,21], ldl.dat[,1])
ivw.object <- ivw_radial(ldl.fdat, 0.05, 1, 0.0001, TRUE)
plotly_radial(ivw.object)
}
\references{
Bowden, J., et al., Improving the visualization, interpretation and analysis of two-sample summary data Mendelian randomization via the Radial plot and Radial regression. International Journal of Epidemiology, 2018. 47(4): p. 1264-1278.
}
\author{
Wes Spiller; Jack Bowden; Tom Palmer.
}
