% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadialMR-package.R
\docType{package}
\name{RadialMR-package}
\alias{RadialMR}
\alias{RadialMR-package}
\title{RadialMR: RadialMR}
\description{
A package for implementing radial inverse variance weighted and MR-Egger methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/WSpiller/RadialMR}
  \item \url{https://wspiller.github.io/RadialMR/}
  \item \url{https://mrcieu.r-universe.dev/RadialMR}
}

}
\author{
\strong{Maintainer}: Tom Palmer \email{remlapmot@hotmail.com} (\href{https://orcid.org/0000-0003-4655-4511}{ORCID})

Authors:
\itemize{
  \item Tom Palmer \email{remlapmot@hotmail.com} (\href{https://orcid.org/0000-0003-4655-4511}{ORCID})
  \item Wes Spiller \email{wes.spiller@bristol.ac.uk} (\href{https://orcid.org/0000-0002-8169-5531}{ORCID})
  \item Jack Bowden \email{j.bowden2@exeter.ac.uk} (\href{https://orcid.org/0000-0003-2628-3304}{ORCID})
}

}
\keyword{internal}
