% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadialMR.R
\name{RadialMR}
\alias{RadialMR}
\title{RadialMR}
\usage{
RadialMR(BXG, BYG, seBXG, seBYG, RSID, METHOD, SCALE, OUTLIERS, ALPHA,
  SCALEMATCH)
}
\arguments{
\item{BXG}{A numeric vector of beta-coefficient values for genetic associations with the first variable (exposure).}

\item{BYG}{A numeric vector of beta-coefficient values for genetic associations with the second variable (outcome).}

\item{seBXG}{The standard errors corresponding to the beta-coefficients BXG.}

\item{seBYG}{The standard errors corresponding to the beta-coefficients BYG.}

\item{RSID}{A vector of names for the genetic variants included in the analysis. If unavailable, \code{RSID="NO"} will generate placeholder ID numbers for each variant.}

\item{METHOD}{Indicates whether to perform radial IVW (\code{"IVW"}),radial MR-Egger (\code{"EGGER"}), or both radial methods (\code{"BOTH"}) simultaneously.}

\item{SCALE}{Indicates whether to produce a plot including a full radial scale (\code{"YES"}), or a scatterplot showing only the effect estimates (\code{"NO"}).}

\item{OUTLIERS}{Indicates whether display only the set of variants identified as outliers (\code{"YES"}) or the complete set of variants (\code{"NO"}). Note that when (\code{"OUTLIERS=YES"}), non-outlying variants further from the origin than the furthest outlier will cause an error message that one or more points have been omitted. These are non-outlying variants beyond the scale.}

\item{ALPHA}{A value specifying the statistical significance threshold for identifying outliers (\code{"0.05"} specifies a p-value threshold of 0.05)}

\item{SCALEMATCH}{Indicates whether x and y axes should have the same range(\code{"YES"}), or different ranges (\code{"NO"}).}
}
\value{
An object containing at least the following components:\describe{
\item{\code{plot}}{A radial plot showing one or more radial MR estimates.}
\item{\code{Q}}{A data frame giving the overall Q statistic for the model method and respective chi-squared test. If \code{METHOD="BOTH"}, separate Q statistics and chi-square values are given for each model fit.}
\item{\code{outliers}}{A data frame containing variants identified as outliers, with respective Q statistics, chi-squared tests and SNP identification. In cases where \code{METHOD="BOTH"}, the dataframe will contain variants which are identified as an outlier for either model.}
\item{\code{IVW}}{A summary output for the radial IVW regression fit. If \code{METHOD="EGGER"}, will return \code{"NULL"}.}
\item{\code{egger}}{A summary output for the radial MR-Egger regression fit If \code{METHOD="IVW"}, will return \code{NULL}.}
}
}
\description{
Fits radial forms of Inverse Variance Weighted (IVW) and MR-Egger regression. It provides output summaries for each method, and plotting options for visualising goodness-of-fit using Cochran's Q statistic. This allows for identification of outlying observations and is coding invariant unlike conventional MR-Egger regression.
}
\examples{

RadialMR(Data$beta.exposure,Data$beta.outcome,Data$se.exposure,Data$se.outcome,Data$SNP,"BOTH","YES","NO",0.05,"NO")$plot

}
\references{
Bowden, J. et al., 2017. "Improving the visualisation, interpretation and analysis of two-sample summary data Mendelian randomization via the radial plot and radial regression." BioRxiv. doi: https://doi.org/10.1101/200378. Available at: \url{https://www.biorxiv.org/content/early/2017/10/11/200378}
}
\author{
Wes Spiller; Jack Bowden.
}
