% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmr_to_rmr_format.R
\name{tsmr_to_rmr_format}
\alias{tsmr_to_rmr_format}
\title{Convert a data.frame containing a single exposure - outcome pair generated by TwoSampleMR::harmonise_data() to the RadialMR rmr_format class}
\usage{
tsmr_to_rmr_format(dat)
}
\arguments{
\item{dat}{Output for a single exposure-outcome pair from \code{\link[TwoSampleMR:harmonise_data]{TwoSampleMR::harmonise_data}}.}
}
\value{
Object of class \code{rmr_format}, the RadialMR format
}
\description{
Creates an object of RadialMR format, i.e. of class \code{rmr_format}, for a single exposure - outcome pair.
}
\details{
Only the rows where the column \code{mr_keep} are \code{TRUE} are kept.
}
\examples{
\dontrun{
# Example with one exposure-outcome pair
bmi_exp_dat <- TwoSampleMR::extract_instruments(outcomes = 'ieu-a-2')
chd_out_dat <- TwoSampleMR::extract_outcome_data(
                               snps = bmi_exp_dat$SNP,
                               outcomes = 'ieu-a-7')
dat <- TwoSampleMR::harmonise_data(exposure_dat = bmi_exp_dat,
                                   outcome_dat = chd_out_dat)
dat <- tsmr_to_rmr_format(dat)
class(dat)
head(dat)
}
}
