% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egger_radial.R
\name{egger_radial}
\alias{egger_radial}
\title{egger_radial}
\usage{
egger_radial(r_input, alpha, weights, summary)
}
\arguments{
\item{r_input}{A formatted data frame using the \code{format_radial} function.}

\item{alpha}{A value specifying the statistical significance threshold for identifying outliers (\code{0.05} specifies a p-value threshold of 0.05).}

\item{weights}{A value specifying the inverse variance weights used to calculate the MR-Egger estimate and Rucker's Q statistic. By default modified second order weights are used, but one can choose to select first order (\code{1}), second order (\code{2}) or modified second order weights (\code{3}).}

\item{summary}{Indicates whether a summary of the radial MR-Egger model should be printed (default= \code{TRUE}) or withheld (\code{FALSE}).}
}
\value{
An object of class \code{"egger"} containing the following components:\describe{
\item{\code{coef}}{A matrix giving the intercept and slope coefficient, corresponding standard errors, t-statistics, and (two-sided) p-values.}
\item{\code{qstatistic}}{Rucker's Q statistic for overall heterogeneity.}
\item{\code{df}}{Degrees of freedom. This is equal to the number of variants -2 when fitting the radial MR-Egger model.}
\item{\code{outliers}}{A data frame containing variants identified as outliers, with respective Q statistics, chi-squared tests and SNP identification.}
\item{\code{data}}{A data frame containing containing SNP IDs, inverse variance weights, the product of the inverse variance weight and ratio estimate for each variant, contribution to overall heterogeneity with corresponding p-value, and a factor indicator showing outlier status.}
\item{\code{confint}}{A vector giving lower and upper confidence limits for the radial MR-Egger effect estimate.}
}
}
\description{
Fits a radial MR-Egger model using first order, second order, or modified second order weights. Outliers are identified using a significance threshold specified by the user. The function returns an object of class \code{"egger"}, containing regression estimates, a measure of total heterogeneity using Rucker's Q statistic, the individual contribution to overall heterogeneity of each variant, and a data frame for use in constructing the radial plot.
}
\examples{

egger_radial(r_input,0.05,1)

}
\references{
Bowden, J. et al., 2017. "Improving the visualisation, interpretation and analysis of two-sample summary data Mendelian randomization via the radial plot and radial regression." BioRxiv. doi: https://doi.org/10.1101/200378. Available at: \url{https://www.biorxiv.org/content/early/2017/10/11/200378}
}
\author{
Wes Spiller; Jack Bowden.
}
