% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egger_radial.R
\name{egger_radial}
\alias{egger_radial}
\title{egger_radial}
\usage{
egger_radial(r_input, alpha, weights, summary)
}
\arguments{
\item{r_input}{A formatted data frame using the \code{\link{format_radial}} function, or an object of class \code{MRInput} generated by \code{\link[MendelianRandomization:mr_input]{MendelianRandomization::mr_input}} or a data.frame for a single exposure-outcome pair generated by \code{\link[TwoSampleMR:harmonise_data]{TwoSampleMR::harmonise_data}}.}

\item{alpha}{A value specifying the statistical significance threshold for identifying outliers (\code{0.05} specifies a p-value threshold of 0.05).}

\item{weights}{A value specifying the inverse variance weights used to calculate the MR-Egger estimate and Rucker's Q statistic. By default modified second order weights are used, but one can choose to select first order (\code{1}), second order (\code{2}) or modified second order weights (\code{3}).}

\item{summary}{A logical argument (\code{TRUE} or \code{FALSE}) indicating whether a summary of results and heterogeneity should be presented (default= \code{TRUE}).}
}
\value{
An object of class \code{"egger"} containing the following components:\describe{
\item{\code{coef}}{A matrix giving the intercept and slope coefficient, corresponding standard errors, t-statistics, and (two-sided) p-values.}
\item{\code{qstatistic}}{Rucker's Q statistic for overall heterogeneity.}
\item{\code{df}}{Degrees of freedom. This is equal to the number of variants -2 when fitting the radial MR-Egger model.}
\item{\code{outliers}}{A data frame containing variants identified as outliers, with respective Q statistics, chi-squared tests and SNP identification.}
\item{\code{data}}{A data frame containing containing SNP IDs, inverse variance weights, the product of the inverse variance weight and ratio estimate for each variant, contribution to overall heterogeneity with corresponding p-value, and a factor indicator showing outlier status.}
\item{\code{confint}}{A vector giving lower and upper confidence limits for the radial MR-Egger effect estimate.}

}
}
\description{
Fits a radial MR-Egger model using first order, second order, or modified second order weights. Outliers are identified using a significance threshold specified by the user. The function returns an object of class \code{"egger"}, containing regression estimates, a measure of total heterogeneity using Rucker's Q statistic, the individual contribution to overall heterogeneity of each variant, and a data frame for use in constructing the radial plot.
}
\examples{
# Example using format_radial data
ldl.dat <- data_radial[data_radial[,10]<5*10^-8,]

ldl.fdat <- format_radial(ldl.dat[,6], ldl.dat[,9],
                          ldl.dat[,15], ldl.dat[,21],
                          ldl.dat[,1])

egger_radial(ldl.fdat, 0.05, 1, TRUE)

# Example using TwoSampleMR format data
\dontrun{
# Example with one exposure-outcome pair
bmi_exp_dat <- TwoSampleMR::extract_instruments(outcomes = 'ieu-a-2')
chd_out_dat <- TwoSampleMR::extract_outcome_data(
                               snps = bmi_exp_dat$SNP,
                               outcomes = 'ieu-a-7')
tsmrdat <- TwoSampleMR::harmonise_data(exposure_dat = bmi_exp_dat,
                                   outcome_dat = chd_out_dat)
egger_radial(r_input = tsmrdat, alpha = 0.05,
             weights = 1, summary = TRUE)
}

# Example using MendelianRandomization format data
dat <- data_radial[data_radial[,10] < 5e-8,]
mrdat <- MendelianRandomization::mr_input(bx = dat$ldlcbeta,
                                          bxse = dat$ldlcse,
                                          by = dat$chdbeta,
                                          byse = dat$chdse,
                                          snps = dat$rsid)
egger_radial(r_input = mrdat, alpha = 0.05,
             weights = 1, summary = TRUE)
}
\references{
Bowden, J., et al., Improving the visualization, interpretation and analysis of two-sample summary data Mendelian randomization via the Radial plot and Radial regression. International Journal of Epidemiology, 2018. 47(4): p. 1264-1278.
}
\author{
Wes Spiller; Jack Bowden; Tom Palmer.
}
