% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_radial.R
\name{format_radial}
\alias{format_radial}
\title{format_radial}
\usage{
format_radial(BXG, BYG, seBXG, seBYG, RSID)
}
\arguments{
\item{BXG}{A numeric vector of beta-coefficient values for genetic associations with the first variable (exposure).}

\item{BYG}{A numeric vector of beta-coefficient values for genetic associations with the second variable (outcome).}

\item{seBXG}{The standard errors corresponding to the beta-coefficients \code{BXG}.}

\item{seBYG}{The standard errors corresponding to the beta-coefficients \code{BYG}.}

\item{RSID}{A vector of names for genetic variants included in the analysis. If variant IDs are not provided (\code{RSID="NULL"}), a vector of ID numbers will be generated.}
}
\value{
A formatted data frame.
}
\description{
Reads in summary data. Checks and organises columns for use in calculating radial IVW and radial MR-Egger estimates. Where variant IDs are not provided, a vector is generated for variant identification.
}
\examples{

format_radial(summarydata[,1],summarydata[,3],summarydata[,2],summarydata[,4])

}
\references{
Bowden, J., et al., Improving the visualization, interpretation and analysis of two-sample summary data Mendelian randomization via the Radial plot and Radial regression. International Journal of Epidemiology, 2018. 47(4): p. 1264-1278.
}
\author{
Wes Spiller; Jack Bowden.
}
